I am some genius myself. Like this wasn't even difficult whole lobby creator with main menu is ready to go. Whole UI is teady, tedy? Tomorrow I will finish Lobby and LobbySearcher UI with their scripts respectably
Giga Niga :D

+++ Szemrany Biznes/Assets/Script/GameLobby.cs
@@ -80,7 +80,7 @@ public class GameLobby : MonoBehaviour
 
 
     
-    public async void CreateLobby(string lobbyName, int maxPlayers, bool IsPrivate) //Tworzy lobby
+    public async void CreateLobby(string lobbyName, float maxPlayers, bool IsPrivate) //Tworzy lobby
     {
         try
         {
@@ -95,7 +95,7 @@ public class GameLobby : MonoBehaviour
                 }*/
             };
 
-            Unity.Services.Lobbies.Models.Lobby lobby = await LobbyService.Instance.CreateLobbyAsync(lobbyName, maxPlayers, createLobbyOptions);
+            Unity.Services.Lobbies.Models.Lobby lobby = await LobbyService.Instance.CreateLobbyAsync(lobbyName, (int)maxPlayers, createLobbyOptions);
 
             hostLobby = lobby;
             joinedLobby = hostLobby;
@@ -179,8 +179,8 @@ public class GameLobby : MonoBehaviour
         return new Player
         {
             Data = new Dictionary<string, PlayerDataObject> {
-                        { "PlayerName", new PlayerDataObject(PlayerDataObject.VisibilityOptions.Member, playerName)}
-                    }
+                { "PlayerName", new PlayerDataObject(PlayerDataObject.VisibilityOptions.Member, playerName)}
+            }
         };
     }
